move test globals & mock commands to conftest
