+++ test/test_automated_install.py
@@ -1,16 +1,13 @@
 from textwrap import dedent
-
-SETUPVARS = {
-    'PIHOLE_INTERFACE': 'eth99',
-    'IPV4_ADDRESS': '1.1.1.1',
-    'IPV6_ADDRESS': 'FE80::240:D0FF:FE48:4672',
-    'PIHOLE_DNS_1': '4.2.2.1',
-    'PIHOLE_DNS_2': '4.2.2.2'
-}
-
-tick_box = "[\x1b[1;32m\xe2\x9c\x93\x1b[0m]".decode("utf-8")
-cross_box = "[\x1b[1;31m\xe2\x9c\x97\x1b[0m]".decode("utf-8")
-info_box = "[i]".decode("utf-8")
+from conftest import (
+    SETUPVARS,
+    tick_box,
+    info_box,
+    cross_box,
+    mock_command,
+    mock_command_2,
+    run_script
+)
 
 
 def test_setupVars_are_sourced_to_global_scope(Pihole):
@@ -632,64 +629,3 @@ def test_IPv6_ULA_GUA_test(Pihole):
     ''')
     expected_stdout = 'Found IPv6 ULA address, using it for blocking IPv6 ads'
     assert expected_stdout in detectPlatform.stdout
-
-
-# Helper functions
-def mock_command(script, args, container):
-    '''
-    Allows for setup of commands we don't really want to have to run for real
-    in unit tests
-    '''
-    full_script_path = '/usr/local/bin/{}'.format(script)
-    mock_script = dedent('''\
-    #!/bin/bash -e
-    echo "\$0 \$@" >> /var/log/{script}
-    case "\$1" in'''.format(script=script))
-    for k, v in args.iteritems():
-        case = dedent('''
-        {arg})
-        echo {res}
-        exit {retcode}
-        ;;'''.format(arg=k, res=v[0], retcode=v[1]))
-        mock_script += case
-    mock_script += dedent('''
-    esac''')
-    container.run('''
-    cat <<EOF> {script}\n{content}\nEOF
-    chmod +x {script}
-    rm -f /var/log/{scriptlog}'''.format(script=full_script_path,
-                                         content=mock_script,
-                                         scriptlog=script))
-
-
-def mock_command_2(script, args, container):
-    '''
-    Allows for setup of commands we don't really want to have to run for real
-    in unit tests
-    '''
-    full_script_path = '/usr/local/bin/{}'.format(script)
-    mock_script = dedent('''\
-    #!/bin/bash -e
-    echo "\$0 \$@" >> /var/log/{script}
-    case "\$1 \$2" in'''.format(script=script))
-    for k, v in args.iteritems():
-        case = dedent('''
-        \"{arg}\")
-        echo \"{res}\"
-        exit {retcode}
-        ;;'''.format(arg=k, res=v[0], retcode=v[1]))
-        mock_script += case
-    mock_script += dedent('''
-    esac''')
-    container.run('''
-    cat <<EOF> {script}\n{content}\nEOF
-    chmod +x {script}
-    rm -f /var/log/{scriptlog}'''.format(script=full_script_path,
-                                         content=mock_script,
-                                         scriptlog=script))
-
-
-def run_script(Pihole, script):
-    result = Pihole.run(script)
-    assert result.rc == 0
-    return result
